requirements.txt
Flask
firebase-admin
gunicorn
flask-cors
python-dotenv
